sum adjacent bins in double precision

+++ spectrum.c
@@ -71,7 +71,7 @@ void *demod_spectrum(void *arg){
   while(downconvert(chan) == 0){
     int binp = 0; 
     for(int i=0; i < chan->spectrum.bin_count; i++){ // For each noncoherent integration bin above center freq
-      float p = 0;
+      double p = 0;
       for(int j=0; j < binsperbin; j++){ // Add energy of each fft bin that's part of this user integration bin
 	p += cnrmf(chan->filter.out.fdomain[binp++]);
       }
